From int to long

+++ Controllers/BusinessController.cs
@@ -36,7 +36,7 @@ namespace _3lab_komanda32.Controllers
 
         // GET api/<BusinessController>/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Business>> Get(int id)
+        public async Task<ActionResult<Business>> Get(long id)
         {
             try
             {
@@ -74,7 +74,7 @@ namespace _3lab_komanda32.Controllers
 
         // PUT api/<BusinessController>/5
         [HttpPut("{id}")]
-        public async Task<ActionResult<Business?>> Put(int id, [FromBody] Business business)
+        public async Task<ActionResult<Business?>> Put(long id, [FromBody] Business business)
         {
             try
             {
@@ -96,7 +96,7 @@ namespace _3lab_komanda32.Controllers
 
         // DELETE api/<BusinessController>/5
         [HttpDelete("{id}")]
-        public async Task<ActionResult<Business?>> Delete(int id)
+        public async Task<ActionResult<Business?>> Delete(long id)
         {
 
             try
@@ -120,7 +120,7 @@ namespace _3lab_komanda32.Controllers
         //addresses
         // GET api/<BusinessController>/5/Address
         [HttpGet("{id}/Address")]
-        public async Task<ActionResult<Address>> GetAddress(int id)
+        public async Task<ActionResult<Address>> GetAddress(long id)
         {
             try
             {
@@ -138,7 +138,7 @@ namespace _3lab_komanda32.Controllers
 
         // PUT api/<BusinessController>/5/Address
         [HttpPut("{id}/Address")]
-        public async Task<ActionResult<Address?>> Put(int id, [FromBody] Address address)
+        public async Task<ActionResult<Address?>> Put(long id, [FromBody] Address address)
         {
             try
             {
@@ -162,7 +162,7 @@ namespace _3lab_komanda32.Controllers
 
         // POST api/<BusinessController>/5/ManagePrivileges
         [HttpPost("{id}/ManagePrivileges")]
-        public async Task<ActionResult<Business>> PostPrivilege(int id, [FromBody] ManagePrivilege managePrivilege)
+        public async Task<ActionResult<Business>> PostPrivilege(long id, [FromBody] ManagePrivilege managePrivilege)
         {
             try
             {
